Remove build_wheels.py
